requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
email-validator==2.1.0.post1
Flask-Admin>=1.6.1
google-cloud-texttospeech==2.15.0
Flask-Babel==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
Flask-Caching==2.1.0
sentry-sdk[flask]==1.39.1
prometheus-flask-exporter==0.22.4
python-json-logger==2.0.7 
openai==1.79.0
Flask-WTF==1.2.1
stripe==12.1.0